+++ tests/test_follow_foreign_keys.py
@@ -2,9 +2,9 @@ from pytest_mock_resources import create_postgres_fixture
 from sqlalchemy import Column, ForeignKey, MetaData, types
 from sqlalchemy.ext.declarative import declarative_base
 
-from databudgie.config.models import RootConfig
-from databudgie.etl.backup import backup_all
-from databudgie.etl.restore import restore_all
+from databudgie.backup import backup_all
+from databudgie.config import RootConfig
+from databudgie.restore import restore_all
 from tests.utils import mock_s3_csv, s3_config
 
 metadata = MetaData()
